+++ dpcomp/CMakeLists.txt
@@ -40,62 +40,62 @@ endif()
 add_subdirectory(include/plier)
 
 set(SOURCES_LIST
-    src/analysis/memory_ssa_analysis.cpp
+    src/Conversion/SCFToAffine/SCFToAffine.cpp
     src/analysis/memory_ssa.cpp
+    src/analysis/memory_ssa_analysis.cpp
     src/compiler/compiler.cpp
     src/compiler/pipeline_registry.cpp
     src/dialect.cpp
-    src/rewrites/arg_lowering.cpp
-    src/rewrites/call_lowering.cpp
-    src/rewrites/canonicalize_reductions.cpp
-    src/rewrites/cast_lowering.cpp
-    src/rewrites/common_opts.cpp
-    src/rewrites/cse.cpp
-    src/rewrites/index_type_propagation.cpp
-    src/rewrites/if_rewrites.cpp
-    src/rewrites/loop_rewrites.cpp
-    src/rewrites/memory_rewrites.cpp
-    src/rewrites/promote_to_parallel.cpp
-    src/rewrites/type_conversion.cpp
-    src/rewrites/uplift_math_calls.cpp
+    src/transforms/arg_lowering.cpp
+    src/transforms/call_lowering.cpp
+    src/transforms/canonicalize_reductions.cpp
+    src/transforms/cast_lowering.cpp
     src/transforms/cast_utils.cpp
+    src/transforms/common_opts.cpp
     src/transforms/const_utils.cpp
+    src/transforms/cse.cpp
     src/transforms/func_utils.cpp
+    src/transforms/if_rewrites.cpp
+    src/transforms/index_type_propagation.cpp
     src/transforms/inline_utils.cpp
+    src/transforms/loop_rewrites.cpp
     src/transforms/loop_utils.cpp
+    src/transforms/memory_rewrites.cpp
     src/transforms/pipeline_utils.cpp
+    src/transforms/promote_to_parallel.cpp
+    src/transforms/type_conversion.cpp
+    src/transforms/uplift_math_calls.cpp
     src/utils.cpp
-    src/Conversion/SCFToAffine/SCFToAffine.cpp
     )
 set(HEADERS_LIST
-    include/plier/analysis/memory_ssa_analysis.hpp
+    include/plier/Conversion/SCFToAffine/SCFToAffine.h
+    include/plier/PlierOps.td
     include/plier/analysis/memory_ssa.hpp
+    include/plier/analysis/memory_ssa_analysis.hpp
     include/plier/compiler/compiler.hpp
     include/plier/compiler/pipeline_registry.hpp
     include/plier/dialect.hpp
-    include/plier/PlierOps.td
-    include/plier/pass/rewrite_wrapper.hpp
-    include/plier/rewrites/arg_lowering.hpp
-    include/plier/rewrites/call_lowering.hpp
-    include/plier/rewrites/canonicalize_reductions.hpp
-    include/plier/rewrites/cast_lowering.hpp
-    include/plier/rewrites/common_opts.hpp
-    include/plier/rewrites/cse.hpp
-    include/plier/rewrites/index_type_propagation.hpp
-    include/plier/rewrites/if_rewrites.hpp
-    include/plier/rewrites/loop_rewrites.hpp
-    include/plier/rewrites/memory_rewrites.hpp
-    include/plier/rewrites/promote_to_parallel.hpp
-    include/plier/rewrites/type_conversion.hpp
-    include/plier/rewrites/uplift_math_calls.hpp
+    include/plier/transforms/arg_lowering.hpp
+    include/plier/transforms/call_lowering.hpp
+    include/plier/transforms/canonicalize_reductions.hpp
+    include/plier/transforms/cast_lowering.hpp
     include/plier/transforms/cast_utils.hpp
+    include/plier/transforms/common_opts.hpp
     include/plier/transforms/const_utils.hpp
+    include/plier/transforms/cse.hpp
     include/plier/transforms/func_utils.hpp
+    include/plier/transforms/if_rewrites.hpp
+    include/plier/transforms/index_type_propagation.hpp
     include/plier/transforms/inline_utils.hpp
+    include/plier/transforms/loop_rewrites.hpp
     include/plier/transforms/loop_utils.hpp
+    include/plier/transforms/memory_rewrites.hpp
     include/plier/transforms/pipeline_utils.hpp
+    include/plier/transforms/promote_to_parallel.hpp
+    include/plier/transforms/rewrite_wrapper.hpp
+    include/plier/transforms/type_conversion.hpp
+    include/plier/transforms/uplift_math_calls.hpp
     include/plier/utils.hpp
-    include/plier/Conversion/SCFToAffine/SCFToAffine.h
     )
 
 set(PLIER_LIB "plier")
