+++ dpcomp/src/transforms/canonicalize_reductions.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "plier/rewrites/canonicalize_reductions.hpp"
+#include "plier/transforms/canonicalize_reductions.hpp"
 
 #include <mlir/Dialect/MemRef/IR/MemRef.h>
 #include <mlir/Dialect/SCF/SCF.h>
