+++ dpcomp/src/transforms/cse.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "plier/rewrites/cse.hpp"
+#include "plier/transforms/cse.hpp"
 
 #include <llvm/ADT/ScopedHashTable.h>
 #include <llvm/Support/Allocator.h>
