+++ dpcomp/src/transforms/loop_rewrites.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "plier/rewrites/loop_rewrites.hpp"
+#include "plier/transforms/loop_rewrites.hpp"
 #include "plier/transforms/const_utils.hpp"
 
 #include <mlir/Dialect/Arithmetic/IR/Arithmetic.h>
