+++ numba_dpcomp/mlir_compiler/src/pipelines/plier_to_linalg.cpp
@@ -50,21 +50,21 @@
 #include "pipelines/pre_low_simplifications.hpp"
 
 #include "plier/Conversion/SCFToAffine/SCFToAffine.h"
-#include "plier/pass/rewrite_wrapper.hpp"
-#include "plier/rewrites/call_lowering.hpp"
-#include "plier/rewrites/canonicalize_reductions.hpp"
-#include "plier/rewrites/common_opts.hpp"
-#include "plier/rewrites/cse.hpp"
-#include "plier/rewrites/loop_rewrites.hpp"
-#include "plier/rewrites/memory_rewrites.hpp"
-#include "plier/rewrites/promote_to_parallel.hpp"
-#include "plier/rewrites/type_conversion.hpp"
-#include "plier/rewrites/uplift_math_calls.hpp"
+#include "plier/transforms/call_lowering.hpp"
+#include "plier/transforms/canonicalize_reductions.hpp"
 #include "plier/transforms/cast_utils.hpp"
+#include "plier/transforms/common_opts.hpp"
 #include "plier/transforms/const_utils.hpp"
+#include "plier/transforms/cse.hpp"
 #include "plier/transforms/inline_utils.hpp"
+#include "plier/transforms/loop_rewrites.hpp"
 #include "plier/transforms/loop_utils.hpp"
+#include "plier/transforms/memory_rewrites.hpp"
 #include "plier/transforms/pipeline_utils.hpp"
+#include "plier/transforms/promote_to_parallel.hpp"
+#include "plier/transforms/rewrite_wrapper.hpp"
+#include "plier/transforms/type_conversion.hpp"
+#include "plier/transforms/uplift_math_calls.hpp"
 
 #include "base_pipeline.hpp"
 #include "loop_utils.hpp"
